Session Token -> Session ID

+++ bridge/auth.go
@@ -2,7 +2,7 @@ package bridge
 
 type ConnectionToken struct {
 	UserID  int64
-	Session uint
+	Session uint64
 }
 
 var ConnectionTokens = make(map[string]ConnectionToken)
@@ -15,7 +15,7 @@ func RemoveToken(token string) {
 	delete(ConnectionTokens, token)
 }
 
-func AddToken(token string, id int64, session uint) {
+func AddToken(token string, id int64, session uint64) {
 	ConnectionTokens[token] = ConnectionToken{
 		UserID:  id,
 		Session: session,
